+++ src/components/index.ts
@@ -1,6 +1,5 @@
 import CategoryList from "./categoryList/CategoryList";
 import PanelButton from "./panelButton/PanelButton";
-import Drawer from "./drawer/Drawer";
 import TaskCountIndicator from "./taskCountIndicator/TaskCountIndicator";
 import ListItem from "./listItem/ListItem";
 import DividerLine from "./dividerLine/DividerLine";
@@ -13,5 +12,4 @@ export {
   TaskCountIndicator,
   CategoryList,
   PanelButton,
-  Drawer,
 };
