+++ src/features/index.ts
@@ -1,3 +1,4 @@
 import SearchBar from "./searchBar/SearchBar";
+import Drawer from "./drawer/Drawer";
 
-export { SearchBar };
+export { Drawer, SearchBar };
