+++ src/pages/dashboard/Dashboard.module.css
@@ -1,3 +1,4 @@
 .dashboard {
   display: flex;
+  padding: 10px;
 }
