+++ src/pages/dashboard/Dashboard.tsx
@@ -1,11 +1,11 @@
 import classes from "./Dashboard.module.css";
-import { Drawer } from "../../components/index";
+import { Drawer } from "../../features/index";
 
 export default function Dashboard() {
   return (
-    <main className="ubuntu-regular">
-      <div className={classes.dashboard}>
-        <Drawer />
+    <main className={`ubuntu-regular ${classes.dashboard}`}>
+      <Drawer />
+      <div>
         <span>Sample</span>
       </div>
     </main>
